+++ src/socio/socio.service.ts
@@ -17,7 +17,8 @@ export class SocioService {
 
     async findOne(id: string): Promise<SocioEntity> {
         const socio: SocioEntity = await this.socioRepository.findOne({where: {id}, relations: ["clubs"] } );
-        
+        if (!socio) 
+            throw new BusinessLogicException("Socio no encontrado", BusinessError.NOT_FOUND)
         return socio;
     }
     
@@ -27,27 +28,30 @@ export class SocioService {
         }
         else
         {
-            throw new BusinessLogicException("The socio's mail has an invalid format", BusinessError.BAD_FORMAT);
+            throw new BusinessLogicException("The socio's mail has an invalid format", BusinessError.PRECONDITION_FAILED);
         }
         
     }
 
     async update(id: string, socio: SocioEntity): Promise<SocioEntity> {
         const persistedSocio: SocioEntity = await this.socioRepository.findOne({where:{id}});
+        if (!socio) 
+            throw new BusinessLogicException("Socio no encontrado", BusinessError.NOT_FOUND)
         if (persistedSocio.email.includes("@"))
             {
                 return await this.socioRepository.save({...persistedSocio, ...socio});
             }
             else
         {
-            throw new BusinessLogicException("The socio's mail has an invalid format", BusinessError.BAD_FORMAT);
+            throw new BusinessLogicException("The socio's mail has an invalid format", BusinessError.PRECONDITION_FAILED);
         }
         
     }
 
     async delete(id: string) {
         const socio: SocioEntity = await this.socioRepository.findOne({where:{id}});
-        
+        if (!socio) 
+            throw new BusinessLogicException("Socio no encontrado", BusinessError.NOT_FOUND)
         await this.socioRepository.remove(socio);
     }
 }
